Initial version of suntimes.sh
This puppy gets the sunrise and sunset from Weather Underground.
You'll need a Wunderground API key to make it work. Place it in
a file suntimes.rc, prepended with API_KEY=

Eventually, this will be used to flick the lights with regards
to the sun. And the whole thing will help me to turn the engine
heater depending on the temperature. But that's for another script.